added data_link class to organize commun code better; changed data_link layer a bit so it is more understandable
