+++ tp03/noncanonical.c
@@ -1,77 +1,6 @@
 /*Non-Canonical Input Processing*/
 #include "noncanonical.h"
 
-static int STOP = FALSE; /* unused */
-
-int readMsg(int fd, unsigned char msg) {
-  int part=0;
-  unsigned char rcv_msg;
-  printf("Reading...\n");
-  while (part!=5) {
-
-    read(fd,&rcv_msg,1);
-    switch (part) {
-      case 0:
-        if(rcv_msg==FLAG){
-          part=1;
-          printf("FLAG: 0x%x\n",rcv_msg);
-        }
-        break;
-      case 1:
-        if(rcv_msg==A){
-          part=2;
-          printf("A: 0x%x\n",rcv_msg);
-        }
-        else {
-          if(rcv_msg==FLAG)
-            part=1;
-          else
-            part=0;
-        }
-        break;
-      case 2:
-        if(rcv_msg==msg){
-          part=3;
-          printf("Control: 0x%x\n",rcv_msg);
-        }
-        else
-          part=0;
-        break;
-      case 3:
-        if(rcv_msg==(A^msg)){
-          part=4;
-          printf("Control BCC: 0x%x\n",rcv_msg);
-        }
-        else
-          part=0;
-        break;
-      case 4:
-        if(rcv_msg==FLAG) {
-          part = 5;
-          printf("FINAL FLAG: 0x%x\nReceived Control\n",rcv_msg);
-        }
-        else
-          part=0;
-        break;
-      default:
-        break;
-    }
-  }
-
-  return TRUE;
-}
-
-void resendMsg(int fd, unsigned char msg) {
-  printf("Resending...\n");
-  unsigned char mesh[5];
-  mesh[0]=FLAG;
-  mesh[1]=A;
-  mesh[2]=msg;
-  mesh[3]=mesh[1]^mesh[2];
-  mesh[4]=FLAG;
-  write(fd,mesh,5);
-}
-
 int reader_main(char* port) {
   int fd;
   struct termios oldtio,newtio;
@@ -118,8 +47,11 @@ int reader_main(char* port) {
 
   printf("New termios structure set\n");
 
-  if(readMsg(fd,SET))
-    resendMsg(fd,UA);
+  if (receive_supervision_frame(fd, SET) == 0) {
+    printf("Sending UA reply...\n");
+    send_supervision_frame(fd, UA);
+  }
+
 
   if(tcsetattr(fd,TCSANOW,&oldtio)==-1){
     perror("tsetattr");
