+++ tp03/writenoncanonical.c
@@ -1,73 +1,16 @@
-/*Non-Canonical Input Processing*/ #include "writenoncanonical.h"
+/*Non-Canonical Input Processing*/
+#include "writenoncanonical.h"
 
 int flag = 1, conta = 1;
 
-static int STOP = FALSE;
+static struct termios oldtio;
+static struct termios newtio;
 
 void atende() { // atende alarme
     printf("alarme # %d\n", conta);
     flag = 1;
     conta++;
 }
-
-void send_supervision_frame(int fd, unsigned char msg) {
-    unsigned char mesh[5];
-    mesh[0] = FLAG;
-    mesh[1] = A;
-    mesh[2] = msg;
-    mesh[3] = mesh[1] ^ mesh[2];
-    mesh[4] = FLAG;
-    write(fd, mesh, 5);
-}
-
-void receiveUA(int* part, unsigned char* msg) {
-    switch (*part) {
-    case 0:
-        if (*msg == FLAG) {
-            *part = 1;
-            printf("FLAG: 0x%x\n", * msg);
-        }
-
-        break;
-    case 1:
-        if (*msg == A) {
-            *part = 2;
-            printf("A: 0x%x\n", * msg);
-        } else {
-            if (*msg == FLAG)
-                *part = 1;
-            else
-                *part = 0;
-        }
-        break;
-    case 2:
-        if (*msg == UA) {
-            *part = 3;
-            printf("UA: 0x%x\n", * msg);
-        } else
-            *part = 0;
-        break;
-    case 3:
-        if (*msg == (UA_BCC)) {
-            *part = 4;
-            printf("UA_BCC: 0x%x\n", * msg);
-        } else
-            *part = 0;
-        break;
-    case 4:
-        if (*msg == FLAG) {
-            STOP = TRUE;
-            printf("FINAL FLAG: 0x%x\nReceived UA\n", * msg);
-        } else
-            *part = 0;
-        break;
-    default:
-        break;
-    }
-}
-
-static struct termios oldtio;
-static struct termios newtio;
 
 int open_writer(char* port) {
 		/* top level layer should verifiy ports name */
@@ -129,18 +72,14 @@ int close_writer(int fd) {
 }
 
 int send_set(int fd) {
-    unsigned char rcv_msg;
-
     do {
         send_supervision_frame(fd, SET);
         alarm(3); // activa alarme de 3s
         printf("Sent mesh\n");
         flag = 0;
-        int part = 0;
         printf("Receiving response...\n");
-        while (!STOP && !flag) {
-            read(fd, &rcv_msg, 1);
-            receiveUA(&part, &rcv_msg);
+        while (1) {
+            if (receive_supervision_frame(fd, UA) == 0) break;
         }
         if (flag) printf("Timed Out - Retrying\n");
     } while (conta < 4 && flag);
