+++ tp03/writenoncanonical.h
@@ -1,17 +1,7 @@
-#include "macros.h"
+#include "data_link.h"
+
 
 int send_set(int fd); /* main function for writer */
-
-/**
-* takes file descriptor (port) and sends a code msg in a
-* supervision frame
-*/
-void send_supervision_frame(int fd, unsigned char msg);
-
-/**
-* function to receive UA message
-*/
-void receiveUA(int* part, unsigned char* msg);
 
 /**
 * function to open and set port
